修复我的钱包Bug、版本更新对话框UI

+++ lib/src/main/java/com/alaer/lib/api/bean/AssetsTotalInfo.java
@@ -12,7 +12,7 @@ public class AssetsTotalInfo {
 
     public float total;
     public List<Assets> assets;
-    public double totalUsdt;
+    public float totalUsdt;
     public String baseCurrency;
     public String base;
